Ported WinQuake to 64-bit, CMake, and added SDL2
Initial check-in, this CL is big enough as it is. May still have some
bugs in the 64-bit port.

+++ WinQuake/cl_parse.c
@@ -147,7 +147,7 @@ void CL_KeepaliveMessage (void)
 {
 	float	time;
 	static float lastmsg;
-	int		ret;
+	size_t		ret;
 	sizebuf_t	old;
 	byte		olddata[8192];
